Fixed vanishing bots list, and separated utils functions.
