+++ packages/app/main/src/utils/botListsAreDifferent.ts
@@ -1,0 +1,60 @@
+//
+// Copyright (c) Microsoft. All rights reserved.
+// Licensed under the MIT license.
+//
+// Microsoft Bot Framework: http://botframework.com
+//
+// Bot Framework Emulator Github:
+// https://github.com/Microsoft/BotFramwork-Emulator
+//
+// Copyright (c) Microsoft Corporation
+// All rights reserved.
+//
+// MIT License:
+// Permission is hereby granted, free of charge, to any person obtaining
+// a copy of this software and associated documentation files (the
+// "Software"), to deal in the Software without restriction, including
+// without limitation the rights to use, copy, modify, merge, publish,
+// distribute, sublicense, and/or sell copies of the Software, and to
+// permit persons to whom the Software is furnished to do so, subject to
+// the following conditions:
+//
+// The above copyright notice and this permission notice shall be
+// included in all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED ""AS IS"", WITHOUT WARRANTY OF ANY KIND,
+// EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
+// MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+// NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE
+// LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
+// OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
+// WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+//
+
+import { BotInfo } from '@bfemulator/app-shared';
+
+/**
+ * Compares two bot lists against each other and returns whether they are different or not
+ * @param list1 First bot list to compare
+ * @param list2 Second bot list to compare
+ */
+export const botListsAreDifferent = (list1: BotInfo[], list2: BotInfo[]): boolean => {
+  if (list1.length !== list2.length) {
+    return true;
+  }
+  for (let i = 0; i < list1.length; i++) {
+    const bot1 = list1[i];
+    const bot2 = list2[i];
+
+    if (bot1.displayName !== bot2.displayName) {
+      return true;
+    }
+    if (bot1.path !== bot2.path) {
+      return true;
+    }
+    if (bot1.secret !== bot2.secret) {
+      return true;
+    }
+  }
+  return false;
+};
